// 05-proyectos/02-tienda-inventario-online/03-nest/02-nest-produccion/src/perfilcliente/dto/create-perfilcliente.dto.ts
export class CreatePerfilclienteDto {
  id: number;

  nombre: string;

  apellido1: string;

  apellido2: string;

  direccion: string;

  telefono: number;

  genero: string;

  fecha: string;

  dataimg: Buffer;

  ext: string;

  correoLoginCliente: string;

  sessionCorreoLoginCliente: number;
}